Better Singleton & Tag editor as a panel instead
